+++ craft_providers/errors.py
@@ -16,6 +16,7 @@
 #
 
 """Craft provider errors."""
+
 import dataclasses
 import shlex
 import subprocess