+++ craft_providers/lxd/remotes.py
@@ -142,12 +142,24 @@ _PROVIDER_BASE_TO_LXD_REMOTE_IMAGE: Dict[Enum, RemoteImage] = {
     ),
     ubuntu.BuilddBaseAlias.NOBLE: RemoteImage(
         image_name="core24",
-        remote_name=BUILDD_DAILY_REMOTE_NAME,
-        remote_address=BUILDD_DAILY_REMOTE_ADDRESS,
+        remote_name=BUILDD_RELEASES_REMOTE_NAME,
+        remote_address=BUILDD_RELEASES_REMOTE_ADDRESS,
         remote_protocol=ProtocolType.SIMPLESTREAMS,
     ),
     ubuntu.BuilddBaseAlias.ORACULAR: RemoteImage(
         image_name="oracular",
+        remote_name=BUILDD_RELEASES_REMOTE_NAME,
+        remote_address=BUILDD_RELEASES_REMOTE_ADDRESS,
+        remote_protocol=ProtocolType.SIMPLESTREAMS,
+    ),
+    ubuntu.BuilddBaseAlias.PLUCKY: RemoteImage(
+        image_name="plucky",
+        remote_name=BUILDD_DAILY_REMOTE_NAME,
+        remote_address=BUILDD_DAILY_REMOTE_ADDRESS,
+        remote_protocol=ProtocolType.SIMPLESTREAMS,
+    ),
+    ubuntu.BuilddBaseAlias.QUESTING: RemoteImage(
+        image_name="questing",
         remote_name=BUILDD_DAILY_REMOTE_NAME,
         remote_address=BUILDD_DAILY_REMOTE_ADDRESS,
         remote_protocol=ProtocolType.SIMPLESTREAMS,