+++ tests/integration/lxd/conftest.py
@@ -16,6 +16,7 @@
 #
 
 """Fixtures for LXD integration tests."""
+
 import os
 import random
 import string
@@ -34,7 +35,6 @@ from craft_providers.lxd.lxd_provider import LXDProvider
 _xfail_bases = {
     ubuntu.BuilddBaseAlias.XENIAL: "Fails to setup snapd (#582)",
     ubuntu.BuilddBaseAlias.ORACULAR: "24.10 fails setup (#598)",
-    ubuntu.BuilddBaseAlias.DEVEL: "24.10 fails setup (#598)",
 }
 """List of bases that are expected to fail and a reason why."""