+++ tests/unit/lxd/test_remotes.py
@@ -146,6 +146,8 @@ def test_add_remote_race_condition_error(fake_remote_image, mock_lxc, logs):
         (ubuntu.BuilddBaseAlias.JAMMY, "core22"),
         (ubuntu.BuilddBaseAlias.NOBLE, "core24"),
         (ubuntu.BuilddBaseAlias.ORACULAR, "oracular"),
+        (ubuntu.BuilddBaseAlias.PLUCKY, "plucky"),
+        (ubuntu.BuilddBaseAlias.QUESTING, "questing"),
         (ubuntu.BuilddBaseAlias.DEVEL, "devel"),
     ],
 )