+++ docs/reference/changelog.rst
@@ -4,12 +4,38 @@ Changelog
 See the `Releases page`_ on GitHub for a complete list of commits that are
 included in each version.
 
-2.2.0 (2025-Jan-16)
+2.3.0 (2025-May-09)
 -------------------
+
+New features:
+
+- Add support for Ubuntu 25.04 (Plucky).
+- Add support for Ubuntu 25.10 (Questing).
+
+Bug fixes:
+
+- Fix a race condition where multiple processes would fail to launch LXD
+  instances at the same time.
+
+2.2.0 (2025-Feb-11)
+-------------------
+
+New features:
+
 - ``hookutil.py`` now available for dependent projects to clean up lxd
   instances.
 - Dependent projects can now disable lxd ``mknod`` interception via
   ``LXDProvider``'s ``__init__``.
+
+Bug fixes:
+
+- Fix a bug where LXD instances would fail during setup, due to a race
+  condition where the setup began before the instance had started.
+
+2.1.1 (2025-Feb-11)
+-------------------
+- Fix a bug where instances with underscores in their names would fail to
+  launch.
 
 2.1.0 (2025-Jan-10)
 -------------------